More character movement updates, added a script to have objects continuously face the camera

+++ PoliticalEspionage/Assets/_Daybrayk/Extensions/Vector3Extensions.cs
@@ -6,19 +6,19 @@ namespace Daybrayk
 {
     public static class Vector3Extensions
     {
-        public static Vector3 WithX(this Vector3 vector, int value)
+        public static Vector3 WithX(this Vector3 vector, float value)
         {
             vector.x = value;
             return vector;
         }
 
-        public static Vector3 WithY(this Vector3 vector, int value)
+        public static Vector3 WithY(this Vector3 vector, float value)
         {
             vector.y = value;
             return vector;
         }
 
-        public static Vector3 WithZ(this Vector3 vector, int value)
+        public static Vector3 WithZ(this Vector3 vector, float value)
         {
             vector.z = value;
             return vector;
